+++ kitchen-sink/vue/src/pages/home.vue
@@ -70,6 +70,12 @@
       <f7-list-item link="/form-storage/" title="Form Storage">
         <f7-icon slot="media" icon="icon-f7"></f7-icon>
       </f7-list-item>
+      <f7-list-item link="/gauge/" title="Gauge">
+        <f7-icon slot="media" icon="icon-f7"></f7-icon>
+      </f7-list-item>
+      <f7-list-item link="/grid/" title="Grid / Layout Grid">
+        <f7-icon slot="media" icon="icon-f7"></f7-icon>
+      </f7-list-item>
       <f7-list-item link="/icons/" title="Icons">
         <f7-icon slot="media" icon="icon-f7"></f7-icon>
       </f7-list-item>
@@ -77,12 +83,6 @@
         <f7-icon slot="media" icon="icon-f7"></f7-icon>
       </f7-list-item>
       <f7-list-item link="/inputs/" title="Inputs">
-        <f7-icon slot="media" icon="icon-f7"></f7-icon>
-      </f7-list-item>
-      <f7-list-item link="/gauge/" title="Gauge">
-        <f7-icon slot="media" icon="icon-f7"></f7-icon>
-      </f7-list-item>
-      <f7-list-item link="/grid/" title="Grid / Layout Grid">
         <f7-icon slot="media" icon="icon-f7"></f7-icon>
       </f7-list-item>
       <f7-list-item link="/lazy-load/" title="Lazy Load">