+++ packages/react/components/page.js
@@ -62,6 +62,13 @@ class F7Page extends React.Component {
     });
   }
 
+  onPagePosition(event) {
+    const position = event.detail.position;
+    this.setState({
+      routerClass: `page-${position}`
+    });
+  }
+
   onPageInit(event) {
     const page = event.detail;
     const {
@@ -274,6 +281,7 @@ class F7Page extends React.Component {
     el.removeEventListener('page:beforeremove', self.onPageBeforeRemove);
     el.removeEventListener('page:stack', self.onPageStack);
     el.removeEventListener('page:unstack', self.onPageUnstack);
+    el.removeEventListener('page:position', self.onPagePosition);
   }
 
   componentDidMount() {
@@ -299,6 +307,7 @@ class F7Page extends React.Component {
     self.onPageBeforeRemove = self.onPageBeforeRemove.bind(self);
     self.onPageStack = self.onPageStack.bind(self);
     self.onPageUnstack = self.onPageUnstack.bind(self);
+    self.onPagePosition = self.onPagePosition.bind(self);
 
     if (ptr) {
       el.addEventListener('ptr:pullstart', self.onPtrPullStart);
@@ -322,6 +331,7 @@ class F7Page extends React.Component {
     el.addEventListener('page:beforeremove', self.onPageBeforeRemove);
     el.addEventListener('page:stack', self.onPageStack);
     el.addEventListener('page:unstack', self.onPageUnstack);
+    el.addEventListener('page:position', self.onPagePosition);
   }
 
   get slots() {